+++ src/view/event/event-offers.js
