+++ src/view/info/info-cost.js
