+++ src/view/info/info-trip.js
